fix: Remove top from outlook requests
The Microosft Graph Delta API has a strange behaviour: it returns only 10 elements per page up to the $top value. This is in contrast with other API endpoints that create pages of $top elements each. So leaving $top in delta requests means up to 250 elements per folder are synchronized (and older messages are never downloaded)